+++ examples/simple_project/variables.tf
@@ -28,6 +28,6 @@ variable "credentials_path" {
 }
 
 variable "default_service_account" {
-  description = "Project default service account setting: can be one of `delete`, `depriviledge`, or `keep`."
+  description = "Project default service account setting: can be one of `delete`, `depriviledge`, `disable`, or `keep`."
 }