+++ helpers/migrate4.py
@@ -14,6 +14,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+# flake8: noqa
+
 import argparse
 import copy
 import subprocess