+++ test/fixtures/fabric_project/variables.tf
@@ -14,30 +14,8 @@
  * limitations under the License.
  */
 
-provider "google" {
-  version = "~> 2.18.1"
+variable "folder_id" {
 }
 
-provider "google-beta" {
-  version = "~> 2.18.1"
+variable "billing_account" {
 }
-
-module "project-factory" {
-  source = "../../../"
-
-  name              = "pf-ci-test-minimal-${var.random_string_for_testing}"
-  random_project_id = true
-  domain            = var.domain
-  org_id            = var.org_id
-  folder_id         = var.folder_id
-  billing_account   = var.billing_account
-
-  activate_apis = [
-    "compute.googleapis.com",
-    "container.googleapis.com",
-  ]
-
-  default_service_account     = "disable"
-  disable_services_on_destroy = "false"
-}
-