+++ test/setup/iam.tf
@@ -59,3 +59,9 @@ resource "google_folder_iam_member" "int_test_folder" {
 resource "google_service_account_key" "int_test" {
   service_account_id = google_service_account.int_test.id
 }
+
+resource "google_billing_account_iam_member" "int_billing_user" {
+  billing_account_id = var.billing_account
+  role               = "roles/billing.user"
+  member             = "serviceAccount:${google_service_account.int_test.email}"
+}