+++ test/setup/main.tf
@@ -20,12 +20,12 @@ resource "random_id" "folder_rand" {
 
 resource "google_folder" "ci_pfactory_folder" {
   display_name = "ci-tests-pfactory-folder-${random_id.folder_rand.hex}"
-  parent       = "folders/${var.folder_id}"
+  parent       = "folders/${replace(var.folder_id, "folders/", "")}"
 }
 
 module "pfactory_project" {
   source  = "terraform-google-modules/project-factory/google"
-  version = "~> 3.0"
+  version = "~> 4.0"
 
   name              = "ci-pfactory-tests"
   random_project_id = true